backend/requirements.txt
# Backend requirements
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
xarray>=2024.3.0
netcdf4>=1.7.0
numpy>=1.26.0
rasterio>=1.3.10
python-multipart>=0.0.9
pydantic-settings>=2.0.0
diskcache>=5.6.0

# Linting & Type Checking
ruff>=0.4.0
black>=24.0.0
mypy>=1.8.0
types-requests

# Testing
pytest>=7.4.0
httpx>=0.27.0
pytest-anyio>=0.0.0
